backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-multipart==0.0.6
psycopg2-binary==2.9.9
passlib==1.7.4
bcrypt==4.1.2
python-dotenv==1.0.1
